Refactor workflow replayer tests
